requirements.txt
Flask==2.1.3
flask-cors==3.0.10
torch>=1.11.0
torchvision>=0.12.0
numpy==1.22.4
pillow==9.1.1
pymongo==4.1.1
requests==2.27.1